Add unit tests for activity and model functionalities; refactor base test case
